Cleaned up sending and replying to messages
Removed 'Verstuur Bericht' button if it was an incoming message.
Removed 'Reageer Bericht' if it is an outgoing message.
Removed 'Bewerk Bericht' if it is an incoming message.
Localized the show message page somewhat more.
Introduced musketeer helper (one for all, all for one)
Added antonym helper.
Refactored the show view spec.