Fixes verifying received messages using client certificate fails because the wrong client certificate appears out of thin air.
Gets the client certificate from the request headers instead of from the unix environment variables which are only set when an instance of passenger is created. When such an instance receives a message and another certificate was already in the environment variables this wrong certificate is used to verify the incoming message.