Added explicit confirmation of sent messages using the message queue
Moved the verification of certificates into the model logic and is now triggered directly from the controller.
Added dependency on delayed job when receiving messages.
Added triggering of a confirmation job when the message is finally created from the reception.
Added confirmation job who's job it is to confirm the message on the other end.
Changed the routes to include the organization_id as a namespace because organization id is used in every communication between antorcha's.
Added sending of the delivery_id to be able to send the confirmation to this resource.
Added a confirmation controller which in turn also checks if the received certificate is ok.
Refactored the verification of certificates from the receptions model into a module which is mixed in into the other models.