Fixed: undefined local variable or method `current_user' for MessageService
Current_user was not delegated to the controller. Current user was introduced in the last sprint to stamp transactions and messages with authors. Added tests to ensure delegation of :current_user and :can? to the controller. Removed logging statement, because logger is also not delegated.

Functional tested the SOAP call. It is time for a SOAP integration test.